Disable CockroachDB for now to return to master
