+++ .github/workflows/integration-tests.yml
@@ -19,7 +19,9 @@ jobs:
       fail-fast: false
       matrix:
         node-version: [16.x, 14.x, 12.x, 10.x]
-        database-type: [postgres, pgnative, mysql, mssql, sqlite3, cockroachdb]
+# Restore after crdb is fully supported
+#        database-type: [postgres, pgnative, mysql, mssql, sqlite3, cockroachdb]
+        database-type: [postgres, pgnative, mysql, mssql, sqlite3]
 
     steps:
       - name: Checkout Repository
