+++ test/integration/logger.js
@@ -5,7 +5,7 @@ const _ = require('lodash');
 const { DRIVER_NAMES } = require('../../lib/constants');
 const { isObject } = require('../../lib/util/is');
 
-const { TEST_TIMESTAMP } = require('../util/constants');
+const { TEST_TIMESTAMP, TEST_ID } = require('../util/constants');
 
 module.exports = function (knex) {
   const client = knex.client;
@@ -92,6 +92,8 @@ module.exports = function (knex) {
         function (memo, val, key) {
           if (_.includes(['created_at', 'updated_at'], key)) {
             memo[key] = TEST_TIMESTAMP;
+          } else if (_.includes(['id'], key)) {
+            memo[key] = TEST_ID;
           } else {
             memo[key] = val;
           }
