+++ test/util/constants.js
@@ -1,8 +1,10 @@
 // Test Constants
 const TEST_TIMESTAMP = new Date();
+const TEST_ID = 'TEST_ID';
 const { DRIVER_NAMES } = require('../../lib/constants');
 
 module.exports = {
+  TEST_ID,
   TEST_TIMESTAMP,
   DRIVER_NAMES,
 };
